HR-Companion/requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
supabase
pypdf
python-dotenv
tiktoken
sentence-transformers
torch